pyproject.toml
[project]
name = "fastapi-web-scraper"
version = "0.1.0"
description = "Fatapi web scraper"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "boto3>=1.35.87",
    "fastapi[standard]>=0.115.6",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.7.0",
    "redis>=5.2.1",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.36",
]

[dependency-groups]
dev = [
    "mypy>=1.14.0",
    "ruff>=0.8.4",
]

[tool.mypy]
strict = true
exclude = ["venv", ".venv"]
